Moved new architecture to main folder
