+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = napari-live-recording
-version = 0.2.1
+version = 0.3.0
 author = Jacopo Abramo
 author_email = jacopo.abramo@gmail.com
 url = https://github.com/jethro33/napari-live-recording
