+++ src/napari_live_recording/ui/widgets.py
@@ -4,13 +4,12 @@ from qtpy.QtWidgets import (
     QWidget, 
     QLabel, 
     QComboBox, 
-    QSpinBox, 
-    QDoubleSpinBox, 
+    QSpinBox,
     QLineEdit, 
     QPushButton
 )
-from superqt import QLabeledSlider
-from qtpy.QtWidgets import QFormLayout, QGridLayout, QGroupBox
+from superqt import QLabeledSlider, QLabeledDoubleSlider
+from qtpy.QtWidgets import QFormLayout, QHBoxLayout, QGridLayout, QGroupBox
 from abc import ABC, abstractmethod
 from dataclasses import replace
 from napari_live_recording.common import ROI
@@ -37,6 +36,7 @@ class LocalWidget(ABC):
             unit (str, optional): parameter unit measure. Defaults to "".
             orientation (str, optional): label orientation on the layout. Defaults to "left".
         """
+        super().__init__()
         self.__name = name
         self.__unit = unit
         labelStr = (self.__name + " (" + self.__unit + ")" if self.__unit != "" else self.__name)
@@ -136,121 +136,9 @@ class ComboBox(LocalWidget):
             "currentIndexChanged" : self.combobox.currentIndexChanged,
             "currentTextChanged" : self.combobox.currentTextChanged
         }
-        
-
-class SpinBox(LocalWidget):
-    def __init__(self, param: Tuple[int, int, int], name: str, unit: str = "", orientation: str = "left") -> None:
-        """SpinBox widget.
-
-        Args:
-            param (Tuple[int, int, int]): parameters for SpinBox settings: (<minimum_value>, <maximum_value>, <starting_value>)
-            name (str): parameter label description.
-            unit (str, optional): parameter unit measure. Defaults to "".
-            orientation (str, optional): label orientation on the layout. Defaults to "left".
-        """
-        self.spinbox = QSpinBox()
-        self.spinbox.lineEdit().setAlignment(Qt.AlignmentFlag.AlignCenter)
-        self.spinbox.setRange(param[0], param[1])
-        self.spinbox.setValue(param[2])
-        super().__init__(self.spinbox, name, unit, orientation)
-    
-    def changeWidgetSettings(self, newParam : Tuple[int, int, int]) -> None:
-        """SpinBox update widget parameter method.
-
-        Args:
-            newParam (Tuple(int, int, int)): new parameters for SpinBox settings: (<minimum_value>, <maximum_value>, <starting_value>)
-        """
-        self.spinbox.setRange(newParam[0], newParam[1])
-        self.spinbox.setValue(newParam[2])
-    
-    @property
-    def value(self) -> int:
-        """Returns the SpinBox current value.
-        """
-        return self.spinbox.value()
-    
-    @value.setter
-    def value(self, value: int) -> None:
-        """Sets the SpinBox current value to show on the widget.
-
-        Args:
-            value (int): value to set.
-        """
-        self.spinbox.setValue(value)
-    
-    @property
-    def signals(self) -> Dict[str, Signal]:
-        """Returns a dictionary of signals available for the SpinBox widget.
-        Exposed signals are:
-        
-        - valueChanged,
-        - textChanged
-
-        Returns:
-            Dict: Dict of signals (key: function name, value: function objects).
-        """
-        return {
-            "valueChanged" : self.spinbox.valueChanged,
-            "textChanged" : self.spinbox.textChanged
-        }
-
-class DoubleSpinBox(LocalWidget):
-    def __init__(self, param: Tuple[float, float, float], name: str, unit: str = "", orientation: str = "left") -> None:
-        """DoubleSpinBox widget.
-
-        Args:
-            param (Tuple[float, float, float]): parameters for spinbox settings: (<minimum_value>, <maximum_value>, <starting_value>)
-            name (str): parameter label description.
-            unit (str, optional): parameter unit measure. Defaults to "".
-            orientation (str, optional): label orientation on the layout. Defaults to "left".
-        """
-        self.__spinbox = QDoubleSpinBox()
-        self.__spinbox.setRange(param[0], param[1])
-        self.__spinbox.setValue(param[2])
-        super().__init__(self.__spinbox, name, unit, orientation)
-
-    def changeWidgetSettings(self, newParam : Tuple[float, float, float]) -> None:
-        """DoubleSpinBox update widget parameter method.
-
-        Args:
-            newParam (Tuple[float, float, float]): new parameters for SpinBox settings: (<minimum_value>, <maximum_value>, <starting_value>)
-        """
-        self.__spinbox.setRange(newParam[0], newParam[1])
-        self.__spinbox.setValue(newParam[2])
-    
-    @property
-    def value(self) -> float:
-        """Returns the DoubleSpinBox current value.
-        """
-        return self.__spinbox.value()
-    
-    @value.setter
-    def value(self, value: float) -> None:
-        """Sets the DoubleSpinBox current value to show on the widget.
-
-        Args:
-            value (float): value to set.
-        """
-        self.__spinbox.setValue(value)
-
-    @property
-    def signals(self) -> Dict[str, Signal]:
-        """Returns a dictionary of signals available for the SpinBox widget.
-        Exposed signals are:
-        
-        - valueChanged,
-        - textChanged
-
-        Returns:
-            Dict: Dict of signals (key: function name, value: function objects).
-        """
-        return {
-            "valueChanged" : self.__spinbox.valueChanged,
-            "textChanged" : self.__spinbox.textChanged
-        }
 
 class LabeledSlider(LocalWidget):
-    def __init__(self, param: Tuple[int, int, int], name: str, unit: str = "", orientation: str = "left") -> None:
+    def __init__(self, param: Union[Tuple[int, int, int], Tuple[float, float, float]], name: str, unit: str = "", orientation: str = "left") -> None:
         """Slider widget.
 
         Args:
@@ -259,7 +147,10 @@ class LabeledSlider(LocalWidget):
             unit (str, optional): parameter unit measure. Defaults to "".
             orientation (str, optional): label orientation on the layout. Defaults to "left".
         """
-        self.__slider = QLabeledSlider(Qt.Horizontal)
+        if any(type(param) == float):
+            self.__slider = QLabeledDoubleSlider(Qt.Horizontal)
+        else:
+            self.__slider = QLabeledSlider(Qt.Horizontal)
         self.__slider.setRange(param[0], param[1])
         self.__slider.setValue(param[2])
         super().__init__(self.__slider, name, unit, orientation)
@@ -425,8 +316,9 @@ class RecordHandling(QObject):
         - album stacking snap.
 
         Widget layout:
-        |(0,0) QPushButton (Snap)|(0,1) QPushButton (Album)|(0,2) QPushButton  (Live) |
-        |(1,0-1)          QSpinBox (Record size)           |(0,2) QPushButton (Record)| 
+        |(0,0) QPushButton (Snap)     |(0,1) QPushButton (Album)| |
+        |(1,0-1)              QPushButton (Live)                  |
+        |(2,0) QSpinBox (Record size) |(2,1) QPushButton (Record) | 
 
         """
         QObject.__init__(self)
@@ -454,9 +346,9 @@ class RecordHandling(QObject):
         
         self.layout.addWidget(self.snap, 0, 0)
         self.layout.addWidget(self.album, 0, 1)
-        self.layout.addWidget(self.live, 0, 2)
-        self.layout.addWidget(self.recordSpinBox, 1, 0, 1, 2)
-        self.layout.addWidget(self.record, 1, 2)
+        self.layout.addWidget(self.live, 1, 0, 1, 2)
+        self.layout.addWidget(self.recordSpinBox, 2, 0)
+        self.layout.addWidget(self.record, 2, 1)
         self.group.setLayout(self.layout)
 
         # whenever the live button is toggled,
@@ -515,25 +407,11 @@ class ROIHandling(QObject):
     def __init__(self, sensorShape : ROI) -> None:
         """ROI Handling widget. Defines a set of non-custom widgets to set the Region Of Interest of the device.
         This widget is common for all devices.
-
-        Args:
-            cameraROI (ROI): data describing the device sensor shape and step value to increment/decrement each parameter.
         """
         QObject.__init__(self)
-        # todo: maybe this is inefficient...
-        # in previous implementation
-        # copying the reference would cause
-        # issues when changing the ROI
-        # so we'll create a local copy
-        # and discard the input
-        self.sensorFullROI = replace(sensorShape)
 
-        # todo: these widgets are not
-        # our custom LocalWidgets
-        # but since they are common
-        # for all types of cameras
-        # it is not worth to customize them...
-        # ... right?
+        self.sensorFullROI = replace(sensorShape)
+        
         self.offsetXLabel = QLabel("Offset X (px)")
         self.offsetXLabel.setAlignment(Qt.AlignmentFlag.AlignCenter)
