--- src/napari_live_recording/widgets/__init__.py
@@ -1,13 +1,0 @@
-from napari_live_recording.widgets.widgets import (
-    Timer,
-    WidgetEnum,
-    LocalWidget,
-    ComboBox,
-    SpinBox,
-    DoubleSpinBox,
-    LabeledSlider,
-    LineEdit,
-    CameraSelection,
-    RecordHandling,
-    ROIHandling
-)
